+++ test/CMakeLists.txt
@@ -48,5 +48,6 @@ endfunction()
 add_subdirectory(core)
 add_subdirectory(fields)
 add_subdirectory(finiteVolume)
+add_subdirectory(linearAlgebra)
 add_subdirectory(mesh)
 add_subdirectory(dsl)
