+++ test/core/CMakeLists.txt
@@ -3,7 +3,6 @@
 
 add_subdirectory(mpi)
 add_subdirectory(primitives)
-add_subdirectory(linearAlgebra)
 
 neofoam_unit_test(dictionary)
 neofoam_unit_test(executor)
